+++ packages/react-dom/src/client/ReactDOMHostConfig.js
@@ -80,7 +80,7 @@ import {
 } from 'shared/ReactFeatureFlags';
 import {HostComponent, HostText} from 'react-reconciler/src/ReactWorkTags';
 import {TOP_BEFORE_BLUR, TOP_AFTER_BLUR} from '../events/DOMTopLevelEventTypes';
-import {listenToReactEvent} from '../events/DOMModernPluginEventSystem';
+import {listenToReactEvent} from '../events/DOMPluginEventSystem';
 
 export type Type = string;
 export type Props = {
