--- packages/react-dom/src/events/SyntheticClipboardEvent.js
@@ -1,22 +1,0 @@
-/**
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-import SyntheticEvent from '../events/SyntheticEvent';
-
-/**
- * @interface Event
- * @see http://www.w3.org/TR/clipboard-apis/
- */
-const SyntheticClipboardEvent = SyntheticEvent.extend({
-  clipboardData: function(event) {
-    return 'clipboardData' in event
-      ? event.clipboardData
-      : window.clipboardData;
-  },
-});
-
-export default SyntheticClipboardEvent;
