--- packages/react-dom/src/events/SyntheticUIEvent.js
@@ -1,15 +1,0 @@
-/**
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-import SyntheticEvent from '../events/SyntheticEvent';
-
-const SyntheticUIEvent = SyntheticEvent.extend({
-  view: null,
-  detail: null,
-});
-
-export default SyntheticUIEvent;
