+++ packages/react-dom/src/events/plugins/BeforeInputEventPlugin.js
@@ -27,9 +27,11 @@ import {
   initialize as FallbackCompositionStateInitialize,
   reset as FallbackCompositionStateReset,
 } from '../FallbackCompositionState';
-import SyntheticCompositionEvent from '../SyntheticCompositionEvent';
-import SyntheticInputEvent from '../SyntheticInputEvent';
-import {accumulateTwoPhaseListeners} from '../DOMModernPluginEventSystem';
+import {
+  SyntheticCompositionEvent,
+  SyntheticInputEvent,
+} from '../SyntheticEvent';
+import {accumulateTwoPhaseListeners} from '../DOMPluginEventSystem';
 
 const END_KEYCODES = [9, 13, 27, 32]; // Tab, Return, Esc, Space
 const START_KEYCODE = 229;
