+++ packages/react-dom/src/events/plugins/ChangeEventPlugin.js
@@ -8,11 +8,11 @@
  */
 import type {AnyNativeEvent} from '../PluginModuleType';
 import type {TopLevelType} from '../TopLevelEventTypes';
-import type {DispatchQueue} from '../DOMModernPluginEventSystem';
+import type {DispatchQueue} from '../DOMPluginEventSystem';
 import type {EventSystemFlags} from '../EventSystemFlags';
 
 import {registerTwoPhaseEvent} from '../EventRegistry';
-import SyntheticEvent from '../SyntheticEvent';
+import {SyntheticEvent} from '../SyntheticEvent';
 import isTextInputElement from '../isTextInputElement';
 import {canUseDOM} from 'shared/ExecutionEnvironment';
 
@@ -38,7 +38,7 @@ import {batchedUpdates} from '../ReactDOMUpdateBatching';
 import {
   processDispatchQueue,
   accumulateTwoPhaseListeners,
-} from '../DOMModernPluginEventSystem';
+} from '../DOMPluginEventSystem';
 
 function registerEvents() {
   registerTwoPhaseEvent('onChange', [
