+++ packages/react-dom/src/events/plugins/SelectEventPlugin.js
@@ -6,7 +6,7 @@
  */
 
 import {canUseDOM} from 'shared/ExecutionEnvironment';
-import SyntheticEvent from '../../events/SyntheticEvent';
+import {SyntheticEvent} from '../../events/SyntheticEvent';
 import isTextInputElement from '../isTextInputElement';
 import shallowEqual from 'shared/shallowEqual';
 
@@ -29,7 +29,7 @@ import {
 } from '../../client/ReactDOMComponentTree';
 import {hasSelectionCapabilities} from '../../client/ReactInputSelection';
 import {DOCUMENT_NODE} from '../../shared/HTMLNodeType';
-import {accumulateTwoPhaseListeners} from '../DOMModernPluginEventSystem';
+import {accumulateTwoPhaseListeners} from '../DOMPluginEventSystem';
 
 const skipSelectionChangeEvent =
   canUseDOM && 'documentMode' in document && document.documentMode <= 11;
