+++ packages/react-dom/src/test-utils/ReactTestUtils.js
@@ -15,7 +15,7 @@ import {
   HostComponent,
   HostText,
 } from 'react-reconciler/src/ReactWorkTags';
-import SyntheticEvent from '../events/SyntheticEvent';
+import {SyntheticEvent} from '../events/SyntheticEvent';
 import invariant from 'shared/invariant';
 import {ELEMENT_NODE} from '../shared/HTMLNodeType';
 import act from './ReactTestUtilsAct';
